chore(Winui3): PR review.
