+++ src/gui4/windows/kDrive client/kDrive client/CustomControls/UpdateExpander.xaml
