+++ src/gui4/windows/kDrive client/kDrive client/ViewModels/Settings/VersionInfo.cs
@@ -1,6 +1,5 @@
 ﻿using Infomaniak.kDrive.Types;
 using System;
-using System.Windows.Media.Animation;
 
 namespace Infomaniak.kDrive.ViewModels
 {
